insta code moved to backend

+++ src/Hooks/Twitter/useConnect.js
@@ -30,17 +30,25 @@ const useConnect = (details) => {
   /* Create popup */
   const openPopup = () => {
     const url = `https://twitter.com/i/oauth2/authorize?response_type=code&client_id=${process.env.REACT_APP_TWITTER_OAUTH2_CLIENT_ID}&redirect_uri=${process.env.REACT_APP_TWITTER_REDIRECT_URI}&scope=tweet.read%20users.read%20follows.read%20offline.access&state=twitter&code_challenge=challenge&code_challenge_method=plain`;
-    const options = "toolbar=no, menubar=no, width=400, height=600";
+    const options = `toolbar=no, menubar=no, width=400, height=550 left=${
+      window.innerWidth / 2 - 200
+    },top=${window.screen.availHeight / 2 - 275}`;
 
-    window.open(url, "twitter", options);
     localStorage.clear("twitterCode");
+    window.open(url, "twitter", options);
+
+    let count = 0;
     let check = setInterval(() => {
       const code = localStorage.getItem("twitterCode");
+      if (count === 20000) {
+        return clearInterval(check);
+      }
       if (code?.length > 0) {
-        clearInterval(check);
         getInfo(code);
         localStorage.clear("twitterCode");
+        return clearInterval(check);
       }
+      count = count + 100;
     }, 100);
   };
